requirement.txt
PyQt5
matplotlib
schedule
sqlite3
pypdf
python-dateutil
